# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vitals Overwatch API - Family Health Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FamilyMemberWearableSource =
  (typeof FamilyMemberWearableSource)[keyof typeof FamilyMemberWearableSource];

export const FamilyMemberWearableSource = {
  apple_watch: "apple_watch",
  oura: "oura",
  whoop: "whoop",
  garmin: "garmin",
  fitbit: "fitbit",
  manual: "manual",
} as const;

export interface FamilyMember {
  id: number;
  name: string;
  age: number;
  relationship: string;
  avatarInitials: string;
  wearableSource: FamilyMemberWearableSource;
  isHighRisk: boolean;
  guardianViewEnabled: boolean;
  createdAt: string;
}

export type CreateFamilyMemberRequestWearableSource =
  (typeof CreateFamilyMemberRequestWearableSource)[keyof typeof CreateFamilyMemberRequestWearableSource];

export const CreateFamilyMemberRequestWearableSource = {
  apple_watch: "apple_watch",
  oura: "oura",
  whoop: "whoop",
  garmin: "garmin",
  fitbit: "fitbit",
  manual: "manual",
} as const;

export interface CreateFamilyMemberRequest {
  name: string;
  age: number;
  relationship: string;
  wearableSource: CreateFamilyMemberRequestWearableSource;
  isHighRisk: boolean;
  guardianViewEnabled: boolean;
}

export interface VitalsReading {
  id: number;
  memberId: number;
  timestamp: string;
  /** HRV in milliseconds */
  heartRateVariability: number;
  /** Resting heart rate in bpm */
  restingHeartRate: number;
  /** Blood oxygen saturation percentage */
  spo2: number;
  sleepScore?: number | null;
  recoveryScore?: number | null;
  /** Body temperature in Celsius */
  bodyTemperature?: number | null;
  respiratoryRate?: number | null;
  /** AI-computed anomaly score (0-1) */
  anomalyScore?: number | null;
  /** List of detected anomaly types */
  anomalyFlags: string[];
}

export interface CreateVitalsRequest {
  timestamp?: string;
  heartRateVariability: number;
  restingHeartRate: number;
  spo2: number;
  sleepScore?: number | null;
  recoveryScore?: number | null;
  bodyTemperature?: number | null;
  respiratoryRate?: number | null;
}

export interface BiometricBaseline {
  memberId: number;
  computedAt: string;
  avgHrv: number;
  stdHrv: number;
  avgRestingHr: number;
  stdRestingHr: number;
  avgSpo2: number;
  stdSpo2: number;
  avgSleepScore?: number | null;
  dataPointCount: number;
  baselineDays: number;
}

export type HealthSummaryStatus =
  (typeof HealthSummaryStatus)[keyof typeof HealthSummaryStatus];

export const HealthSummaryStatus = {
  excellent: "excellent",
  good: "good",
  fair: "fair",
  warning: "warning",
  critical: "critical",
} as const;

export type HealthSummaryTrendHrv =
  (typeof HealthSummaryTrendHrv)[keyof typeof HealthSummaryTrendHrv];

export const HealthSummaryTrendHrv = {
  improving: "improving",
  stable: "stable",
  declining: "declining",
} as const;

export type HealthSummaryTrendHr =
  (typeof HealthSummaryTrendHr)[keyof typeof HealthSummaryTrendHr];

export const HealthSummaryTrendHr = {
  improving: "improving",
  stable: "stable",
  declining: "declining",
} as const;

export interface HealthSummary {
  memberId: number;
  latestReading?: VitalsReading;
  baseline?: BiometricBaseline;
  status: HealthSummaryStatus;
  trendHrv: HealthSummaryTrendHrv;
  trendHr: HealthSummaryTrendHr;
  activeAlerts: number;
}

export type HealthAlertSeverity =
  (typeof HealthAlertSeverity)[keyof typeof HealthAlertSeverity];

export const HealthAlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type HealthAlertAlertType =
  (typeof HealthAlertAlertType)[keyof typeof HealthAlertAlertType];

export const HealthAlertAlertType = {
  hrv_drop: "hrv_drop",
  elevated_hr: "elevated_hr",
  low_spo2: "low_spo2",
  combined_stress: "combined_stress",
  sleep_disruption: "sleep_disruption",
  temperature_spike: "temperature_spike",
  recovery_deficit: "recovery_deficit",
} as const;

export interface HealthAlert {
  id: number;
  memberId: number;
  memberName: string;
  severity: HealthAlertSeverity;
  alertType: HealthAlertAlertType;
  title: string;
  description: string;
  deviationPercent: number;
  detectedAt: string;
  resolved: boolean;
  resolvedAt?: string | null;
}

export type BaselineReportRecentTrends = {
  hrvTrend: string;
  hrTrend: string;
  spo2Trend: string;
  overallAssessment: string;
};

export interface BaselineReport {
  memberId: number;
  memberName: string;
  memberAge: number;
  reportGeneratedAt: string;
  reportPeriodDays: number;
  wearableSource: string;
  baseline?: BiometricBaseline;
  recentTrends: BaselineReportRecentTrends;
  activeAlerts: HealthAlert[];
  vitalsHistory: VitalsReading[];
  clinicalNotes: string;
}

export interface MemberHealthCard {
  member: FamilyMember;
  summary: HealthSummary;
  latestReading?: VitalsReading | null;
  activeAlertCount: number;
}

export interface DashboardOverview {
  totalMembers: number;
  membersWithAlerts: number;
  criticalAlerts: number;
  familyCards: MemberHealthCard[];
  recentAlerts: HealthAlert[];
}

export type GetMemberVitalsParams = {
  days?: number;
};

export type ListAlertsParams = {
  memberId?: number;
  resolved?: boolean;
};

export type GenerateBaselineReportParams = {
  days?: number;
};