# lib/api-zod/src/generated/types/createFamilyMemberRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vitals Overwatch API - Family Health Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFamilyMemberRequestWearableSource } from "./createFamilyMemberRequestWearableSource";

export interface CreateFamilyMemberRequest {
  name: string;
  age: number;
  relationship: string;
  wearableSource: CreateFamilyMemberRequestWearableSource;
  isHighRisk: boolean;
  guardianViewEnabled: boolean;
}