# lib/api-zod/src/generated/types/healthAlert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vitals Overwatch API - Family Health Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { HealthAlertAlertType } from "./healthAlertAlertType";
import type { HealthAlertSeverity } from "./healthAlertSeverity";

export interface HealthAlert {
  id: number;
  memberId: number;
  memberName: string;
  severity: HealthAlertSeverity;
  alertType: HealthAlertAlertType;
  title: string;
  description: string;
  deviationPercent: number;
  detectedAt: Date;
  resolved: boolean;
  resolvedAt?: Date | null;
}