# lib/api-zod/src/generated/types/healthSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vitals Overwatch API - Family Health Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { BiometricBaseline } from "./biometricBaseline";
import type { HealthSummaryStatus } from "./healthSummaryStatus";
import type { HealthSummaryTrendHr } from "./healthSummaryTrendHr";
import type { HealthSummaryTrendHrv } from "./healthSummaryTrendHrv";
import type { VitalsReading } from "./vitalsReading";

export interface HealthSummary {
  memberId: number;
  latestReading?: VitalsReading;
  baseline?: BiometricBaseline;
  status: HealthSummaryStatus;
  trendHrv: HealthSummaryTrendHrv;
  trendHr: HealthSummaryTrendHr;
  activeAlerts: number;
}